+++ app/controllers/questions_controller.rb
@@ -33,9 +33,8 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    return unless current_user == question.author
+    question.destroy if current_user.author_of?(question)
 
-    question.destroy
     redirect_to questions_path
   end
